Input Updates
Code was updated to take in input from the text.txt file, this allows
for longer peices of texts to be inserted without having to worry about
formatting.

On another note, we must make sure that the peice of text is not too
long, as this seems to not be effective, as the possible 1,2,3 letter
words increases. Similar problem to using 4 letter words.